# Use GroupMemberSet.Merge to reduce CPU usage and memory footprint