--- a/pkg/controller/networkpolicy/store/appliedtogroup.go
+++ b/pkg/controller/networkpolicy/store/appliedtogroup.go
@@ -76,11 +76,11 @@ func (event *appliedToGroupEvent) ToWatchEvent(selectors *storage.Selectors, isI
 		} else {
 			currMembers = controlplane.GroupMemberSet{}
 			for _, members := range event.CurrGroup.GroupMemberByNode {
-				currMembers = currMembers.Union(members)
+				currMembers.Merge(members)
 			}
 			prevMembers = controlplane.GroupMemberSet{}
 			for _, members := range event.PrevGroup.GroupMemberByNode {
-				prevMembers = prevMembers.Union(members)
+				prevMembers.Merge(members)
 			}
 		}
 		for _, member := range currMembers.Difference(prevMembers) {