When getting the union of multiple AddressGroups or AppliedToGroups,
agent used the GroupMemberSet.Union method, which always creates an
intermediate set and copies the original items when merging another
set. This wasted significant time and memory when the groups had
massive members. This patch adds the GroupMemberSet.Merge method which
avoids the above overhead.

benchmark comparison when merging two groups with 10K members and 100
members:

```
name                            old time/op    new time/op    delta
RuleCacheUnionAddressGroups-48    4.27ms ±15%    2.15ms ±15%  -49.72%  (p=0.008 n=5+5)

name                            old alloc/op   new alloc/op   delta
RuleCacheUnionAddressGroups-48    1.95MB ± 0%    0.97MB ± 0%  -50.00%  (p=0.008 n=5+5)

name                            old allocs/op  new allocs/op  delta
RuleCacheUnionAddressGroups-48       565 ± 0%       282 ± 0%  -50.16%  (p=0.000 n=4+5)
```

Signed-off-by: Quan Tian <qtian@vmware.com>

For #2457