@antoninbas 2021-07-26
> Why do we not simply re-write `Union`? Is there any usage where the original GroupMemberSets must be perserved?

Union and merge mean different things IMO, so we shouldn't just change the implementation of union