@tnqn 2021-07-27
> > Why do we not simply re-write `Union`? Is there any usage where the original GroupMemberSets must be perserved?
> 
> Union and merge mean different things IMO, so we shouldn't just change the implementation of union

Yes, also Union is the original method of K8s util package. It's widely used in the scenario where constructing a new set is required.