@Dyanngg 2021-07-27
> Yes, also Union is the original method of K8s util package. It's widely used in the scenario where constructing a new set is required.

Thanks for the explanations. LGTM